rename ResolveKind to ValueConstraint
