+++ lib/remap-lang/src/expression.rs
@@ -1,5 +1,4 @@
-use crate::{CompilerState, Object, Result, State, Value, ValueKind};
-use std::fmt;
+use crate::{CompilerState, Object, Result, State, Value, ValueConstraint, ValueKind};
 
 pub(super) mod arithmetic;
 pub(super) mod assignment;
@@ -48,142 +47,9 @@ pub enum Error {
     Variable(#[from] variable::Error),
 }
 
-/// What kind of value an expression is going to resolve to.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum ResolveKind {
-    /// The expression can resolve to any value at runtime.
-    ///
-    /// This applies to any [`Object`] value that hasn't been coerced yet.
-    Any,
-
-    /// The expressions resolves to one of the defined [`ValueKind`]s.
-    OneOf(Vec<ValueKind>),
-
-    /// The expression resolves to an exact value.
-    Exact(ValueKind),
-
-    /// If the expression succeeds, it might resolve to a value, but doesn't
-    /// have to.
-    ///
-    /// For example, and if-statement without an else-condition can resolve to
-    /// nothing if the if-condition does not match.
-    Maybe(Box<ResolveKind>),
-}
-
-impl ResolveKind {
-    /// Returns `true` if this is a [`ResolveKind::Exact`].
-    pub fn is_exact(&self) -> bool {
-        matches!(self, ResolveKind::Exact(_))
-    }
-
-    /// Returns `true` if this is a [`ResolveKind::Any`].
-    pub fn is_any(&self) -> bool {
-        matches!(self, ResolveKind::Any)
-    }
-
-    /// Returns `true` if this is a [`ResolveKind::Maybe`].
-    pub fn is_maybe(&self) -> bool {
-        matches!(self, ResolveKind::Maybe(_))
-    }
-
-    /// Get a collection of [`ValueKind`]s accepted by this [`ResolveKind`].
-    pub fn value_kinds(&self) -> Vec<ValueKind> {
-        use ResolveKind::*;
-
-        match self {
-            Any => ValueKind::all(),
-            OneOf(v) => v.clone(),
-            Exact(v) => vec![*v],
-            Maybe(v) => v.value_kinds(),
-        }
-    }
-
-    /// Merge two [`ResolveKind`]s, such that the new `ResolveKind` provides the
-    /// most constraint possible resolve kind variant.
-    pub fn merge(&self, other: &Self) -> Self {
-        use ResolveKind::*;
-
-        if let Maybe(kind) = self {
-            let other = match other {
-                Maybe(v) => v,
-                _ => other,
-            };
-
-            return Maybe(Box::new(kind.merge(other)));
-        }
-
-        if let Maybe(kind) = other {
-            return Maybe(Box::new(self.merge(kind)));
-        }
-
-        if self.is_any() || other.is_any() {
-            return Any;
-        }
-
-        let mut kinds: Vec<_> = self
-            .value_kinds()
-            .into_iter()
-            .chain(other.value_kinds().into_iter())
-            .collect();
-
-        kinds.sort();
-        kinds.dedup();
-
-        if kinds.len() == 1 {
-            Exact(kinds[0])
-        } else {
-            OneOf(kinds)
-        }
-    }
-
-    /// Returns `true` if the _other_ [`ResolveKind`] is contained within the
-    /// current one.
-    ///
-    /// That is to say, its constraints must be more strict or equal to the
-    /// constraints of the current one.
-    pub fn contains(&self, other: &Self) -> bool {
-        // If we don't expect none, but the other does, the other's requirement
-        // is less strict than ours.
-        if !self.is_maybe() && other.is_maybe() {
-            return false;
-        }
-
-        let self_kinds = self.value_kinds();
-        let other_kinds = other.value_kinds();
-
-        for kind in other_kinds {
-            if !self_kinds.contains(&kind) {
-                return false;
-            }
-        }
-
-        true
-    }
-}
-
-impl fmt::Display for ResolveKind {
-    /// Print a human readable version of the resolve kind constraints.
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use ResolveKind::*;
-
-        match self {
-            Any => f.write_str("any value"),
-            OneOf(v) => {
-                f.write_str("any of ")?;
-                f.write_str(&v.iter().map(|v| v.as_str()).collect::<Vec<_>>().join(", "))
-            }
-            Exact(v) => f.write_str(&v),
-            Maybe(v) => {
-                f.write_str("none or ")?;
-                f.write_str(&v.to_string())
-            }
-        }
-    }
-}
-
 pub trait Expression: Send + Sync + std::fmt::Debug + dyn_clone::DynClone {
     fn execute(&self, state: &mut State, object: &mut dyn Object) -> Result<Option<Value>>;
-    fn resolves_to(&self, state: &CompilerState) -> ResolveKind;
+    fn resolves_to(&self, state: &CompilerState) -> ValueConstraint;
 }
 
 dyn_clone::clone_trait_object!(Expression);
@@ -212,7 +78,7 @@ macro_rules! expression_dispatch {
                 }
             }
 
-            fn resolves_to(&self, state: &CompilerState) -> ResolveKind {
+            fn resolves_to(&self, state: &CompilerState) -> ValueConstraint {
                 match self {
                     $(Expr::$expr(expression) => expression.resolves_to(state)),+
                 }
@@ -248,7 +114,7 @@ mod tests {
 
     #[test]
     fn test_contains() {
-        use ResolveKind::*;
+        use ValueConstraint::*;
         use ValueKind::*;
 
         let cases = vec![
@@ -273,7 +139,7 @@ mod tests {
 
     #[test]
     fn test_merge() {
-        use ResolveKind::*;
+        use ValueConstraint::*;
         use ValueKind::*;
 
         let cases = vec![
