+++ lib/remap-lang/src/expression/arithmetic.rs
@@ -1,5 +1,5 @@
 use super::{
-    CompilerState, Expr, Expression, Object, ResolveKind, Result, State, Value, ValueKind,
+    CompilerState, Expr, Expression, Object, ValueConstraint, Result, State, Value, ValueKind,
 };
 use crate::Operator;
 
@@ -48,23 +48,23 @@ impl Expression for Arithmetic {
         result.map(Some).map_err(Into::into)
     }
 
-    fn resolves_to(&self, state: &CompilerState) -> ResolveKind {
+    fn resolves_to(&self, state: &CompilerState) -> ValueConstraint {
         let lhs_kind = self.lhs.resolves_to(state);
         let rhs_kind = self.rhs.resolves_to(state);
 
         use Operator::*;
         match self.op {
             Or => lhs_kind.merge(&rhs_kind),
-            Multiply | Add => ResolveKind::OneOf(vec![
+            Multiply | Add => ValueConstraint::OneOf(vec![
                 ValueKind::String,
                 ValueKind::Integer,
                 ValueKind::Float,
             ]),
             Remainder | Subtract | Divide => {
-                ResolveKind::OneOf(vec![ValueKind::Integer, ValueKind::Float])
+                ValueConstraint::OneOf(vec![ValueKind::Integer, ValueKind::Float])
             }
             And | Equal | NotEqual | Greater | GreaterOrEqual | Less | LessOrEqual => {
-                ResolveKind::Exact(ValueKind::Boolean)
+                ValueConstraint::Exact(ValueKind::Boolean)
             }
         }
     }
