+++ lib/remap-lang/src/expression/block.rs
@@ -1,4 +1,4 @@
-use crate::{CompilerState, Expr, Expression, Object, ResolveKind, Result, State, Value};
+use crate::{CompilerState, Expr, Expression, Object, ValueConstraint, Result, State, Value};
 
 #[derive(Debug, Clone)]
 pub(crate) struct Block {
@@ -22,10 +22,10 @@ impl Expression for Block {
         Ok(value)
     }
 
-    fn resolves_to(&self, state: &CompilerState) -> ResolveKind {
+    fn resolves_to(&self, state: &CompilerState) -> ValueConstraint {
         self.expressions
             .last()
             .map(|e| e.resolves_to(state))
-            .unwrap_or(ResolveKind::Any)
+            .unwrap_or(ValueConstraint::Any)
     }
 }
