+++ lib/remap-lang/src/expression/if_statement.rs
@@ -1,6 +1,6 @@
 use super::Error as E;
 use crate::{
-    value, CompilerState, Expr, Expression, Object, ResolveKind, Result, State, Value, ValueKind,
+    value, CompilerState, Expr, Expression, Object, ValueConstraint, Result, State, Value, ValueKind,
 };
 
 #[derive(thiserror::Error, Debug, PartialEq)]
@@ -43,12 +43,12 @@ impl Expression for IfStatement {
         }
     }
 
-    fn resolves_to(&self, state: &CompilerState) -> ResolveKind {
+    fn resolves_to(&self, state: &CompilerState) -> ValueConstraint {
         let true_resolves = self.true_expression.resolves_to(state);
         let false_resolves = self.false_expression.resolves_to(state);
 
         if true_resolves.is_any() || false_resolves.is_any() {
-            return ResolveKind::Any;
+            return ValueConstraint::Any;
         }
 
         let true_value_kinds = true_resolves.value_kinds();
@@ -60,15 +60,15 @@ impl Expression for IfStatement {
             .collect::<Vec<_>>();
 
         let resolve = if kinds.len() == 1 {
-            ResolveKind::Exact(kinds[0])
+            ValueConstraint::Exact(kinds[0])
         } else {
-            ResolveKind::OneOf(kinds)
+            ValueConstraint::OneOf(kinds)
         };
 
         // FIXME: at this point, we might only have resolves from if or else, so
         // if the other is "maybe", the result can still be exact.
         if true_resolves.is_maybe() || false_resolves.is_maybe() {
-            ResolveKind::Maybe(Box::new(resolve))
+            ValueConstraint::Maybe(Box::new(resolve))
         } else {
             resolve
         }
