+++ lib/remap-lang/src/expression/noop.rs
@@ -1,4 +1,4 @@
-use crate::{CompilerState, Expression, Object, ResolveKind, Result, State, Value};
+use crate::{CompilerState, Expression, Object, ValueConstraint, Result, State, Value};
 
 #[derive(Debug, Clone)]
 pub struct Noop;
@@ -8,7 +8,7 @@ impl Expression for Noop {
         Ok(None)
     }
 
-    fn resolves_to(&self, _: &CompilerState) -> ResolveKind {
-        ResolveKind::Maybe(Box::new(ResolveKind::Any))
+    fn resolves_to(&self, _: &CompilerState) -> ValueConstraint {
+        ValueConstraint::Maybe(Box::new(ValueConstraint::Any))
     }
 }
